[ADD] stock_release_channel_batch_mode_commercial_partner
This new addon defines the 'Grouped by Commercial Entiry' mode when releasing in batch. This mode was previously defined into the stock_release_channel addon but since it requires to store an additional info on the stock.picking model we moved this functionality in a dedicated addon. This will avoid to create useless additional column on the stock.picking model if we don't need it. Moreover, since having the commercial partner on the stock.picking could make sense outside of the stock_release_channel context, a new addon 'stock_picking_commercial_partner' has been created into the oca/stock-logistics-warehouse repository to provides this generic functionality
