+++ stock_release_channel/tests/test_channel_release_batch.py
@@ -1,7 +1,7 @@
 # Copyright 2020 Camptocamp (https://www.camptocamp.com)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
-from odoo import _, exceptions
+from odoo import exceptions
 
 from .common import ChannelReleaseCase
 
@@ -58,30 +58,3 @@ class TestChannelReleaseBatch(ChannelReleaseCase):
         self.pickings.need_release = False  # cheat for getting the right condition
         action = self.channel.release_next_batch()
         self._assert_action_nothing_in_the_queue(action)
-
-    def test_release_auto_group_commercial_partner(self):
-        self.channel.batch_mode = "group_commercial_partner"
-        self.channel.release_next_batch()
-        self.assertFalse(self.picking.need_release)
-        self.assertFalse(self.picking2.need_release)
-        other_pickings = self.pickings - (self.picking | self.picking2)
-        self.assertTrue(all(p.need_release) for p in other_pickings)
-
-    def test_release_auto_group_commercial_partner_no_next_batch(self):
-        self.channel.batch_mode = "group_commercial_partner"
-        self.pickings.need_release = False  # cheat for getting the right condition
-        action = self.channel.release_next_batch()
-        self._assert_action_nothing_in_the_queue(action)
-
-    def _assert_action_nothing_in_the_queue(self, action):
-        self.assertEqual(
-            action,
-            {
-                "effect": {
-                    "fadeout": "fast",
-                    "message": _("Nothing in the queue!"),
-                    "img_url": "/web/static/src/img/smile.svg",
-                    "type": "rainbow_man",
-                }
-            },
-        )
